03_Final/Code/FootDetection/requirements.txt
numpy
opencv-python-headless
Pillow
scikit-learn
matplotlib
streamlit